add Levinson24 macropad keyboard, dervied from half a Levinson
make levinson24 the same layout on both halves

default was to mirror the layout based on which side of the board was
plugged in. that's not relevant in macropad use, so set board side bit
in EEPROM and use EE_HANDS setting to make the layout work the same way
regardless of which board and which orientation is plugged in.