QMK Configurator update for the Pearl 40% (#2810)
* QMK Configurator updates for the Pearl 40%

Attempt to get the physical layout as displayed in the Configurator more true-to-life.

* Bugfixes per mechmerlin

"By changing KEYMAP to LAYOUT in the .h file, all the keymaps who rely on KEYMAP are now broken. You need to go into the keymap directory and fix all the keymaps affected by this change. Should just be an issue of renaming KEYMAP to LAYOUT."

* Merge pull request #2 from noroadsleft/noroadsleft-patch-20180425

Bugfixes per mechmerlin

* keymap.c updates for Pearl

-#include "pearl.h"
+#QMK_KEYBOARD_H